Consolidate article routes and presentation
